Fix types file & dir
